backend/pyproject.toml
[tool.poetry]
name = "brickshift"
version = "0.1.0"
description = ""
authors = ["Kerem Zopcuk <keremzopcuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.7"
icalendar = "^5.0.10"
django-ical = "^1.9.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
pillow = "^10.1.0"
azure-cognitiveservices-vision-customvision = "^3.1.0"
django-environ = "^0.11.2"
black = "^23.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
